Final_Submission_With_Images/requirements.txt
fastapi
uvicorn
facenet-pytorch
torch
torchvision
Pillow
requests
beautifulsoup4
numpy
